Change the way GLTFDocumentExtension classes are registered
Also move GLTFDocumentExtension into the extensions folder
