+++ modules/gltf/doc_classes/GLTFDocumentExtension.xml
@@ -5,6 +5,8 @@
 	</brief_description>
 	<description>
 		Extends the functionality of the [GLTFDocument] class by allowing you to run arbitrary code at various stages of GLTF import or export.
+		To use, make a new class extending GLTFDocumentExtension, override any methods you need, make an instance of your class, and register it using [method GLTFDocument.register_gltf_document_extension].
+		[b]Note:[/b] Like GLTFDocument itself, all GLTFDocumentExtension classes must be stateless in order to function properly. If you need to store data, use the [code]set_additional_data[/code] and [code]get_additional_data[/code] methods in [GLTFState] or [GLTFNode].
 	</description>
 	<tutorials>
 	</tutorials>
@@ -61,7 +63,9 @@
 		<method name="_import_preflight" qualifiers="virtual">
 			<return type="int" />
 			<param index="0" name="state" type="GLTFState" />
+			<param index="1" name="extensions" type="PackedStringArray" />
 			<description>
+				This callback is run first. It is used to determine if this GLTFDocumentExtension class should be used for importing a given GLTF file. If [constant OK], the import will use this GLTFDocumentExtension class.
 			</description>
 		</method>
 	</methods>
