+++ modules/gltf/extensions/gltf_document_extension.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  gltf_document_extension_convert_importer_mesh.cpp                    */
+/*  gltf_document_extension.h                                            */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,54 +28,34 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#include "gltf_document_extension_convert_importer_mesh.h"
+#ifndef GLTF_DOCUMENT_EXTENSION_H
+#define GLTF_DOCUMENT_EXTENSION_H
 
-#include "gltf_state.h"
+#include "../gltf_state.h"
 
-#include "core/error/error_macros.h"
-#include "scene/3d/mesh_instance_3d.h"
-#include "scene/resources/importer_mesh.h"
+class GLTFDocumentExtension : public Resource {
+	GDCLASS(GLTFDocumentExtension, Resource);
 
-void GLTFDocumentExtensionConvertImporterMesh::_bind_methods() {
-}
+protected:
+	static void _bind_methods();
 
-Error GLTFDocumentExtensionConvertImporterMesh::import_post(Ref<GLTFState> p_state, Node *p_root) {
-	ERR_FAIL_NULL_V(p_root, ERR_INVALID_PARAMETER);
-	ERR_FAIL_NULL_V(p_state, ERR_INVALID_PARAMETER);
-	List<Node *> queue;
-	queue.push_back(p_root);
-	List<Node *> delete_queue;
-	while (!queue.is_empty()) {
-		List<Node *>::Element *E = queue.front();
-		Node *node = E->get();
-		ImporterMeshInstance3D *mesh_3d = cast_to<ImporterMeshInstance3D>(node);
-		if (mesh_3d) {
-			MeshInstance3D *mesh_instance_node_3d = memnew(MeshInstance3D);
-			Ref<ImporterMesh> mesh = mesh_3d->get_mesh();
-			if (mesh.is_valid()) {
-				Ref<ArrayMesh> array_mesh = mesh->get_mesh();
-				mesh_instance_node_3d->set_name(node->get_name());
-				mesh_instance_node_3d->set_transform(mesh_3d->get_transform());
-				mesh_instance_node_3d->set_mesh(array_mesh);
-				mesh_instance_node_3d->set_skin(mesh_3d->get_skin());
-				mesh_instance_node_3d->set_skeleton_path(mesh_3d->get_skeleton_path());
-				node->replace_by(mesh_instance_node_3d);
-				delete_queue.push_back(node);
-			} else {
-				memdelete(mesh_instance_node_3d);
-			}
-		}
-		int child_count = node->get_child_count();
-		for (int i = 0; i < child_count; i++) {
-			queue.push_back(node->get_child(i));
-		}
-		queue.pop_front();
-	}
-	while (!queue.is_empty()) {
-		List<Node *>::Element *E = delete_queue.front();
-		Node *node = E->get();
-		memdelete(node);
-		delete_queue.pop_front();
-	}
-	return OK;
-}
+public:
+	virtual Vector<String> get_supported_extensions();
+	virtual Error import_preflight(Ref<GLTFState> p_state, Vector<String> p_extensions);
+	virtual Error import_post_parse(Ref<GLTFState> p_state);
+	virtual Error export_post(Ref<GLTFState> p_state);
+	virtual Error import_post(Ref<GLTFState> p_state, Node *p_node);
+	virtual Error export_preflight(Node *p_state);
+	virtual Error import_node(Ref<GLTFState> p_state, Ref<GLTFNode> p_gltf_node, Dictionary &r_json, Node *p_node);
+	virtual Error export_node(Ref<GLTFState> p_state, Ref<GLTFNode> p_gltf_node, Dictionary &r_json, Node *p_node);
+	GDVIRTUAL0R(Vector<String>, _get_supported_extensions);
+	GDVIRTUAL2R(int, _import_preflight, Ref<GLTFState>, Vector<String>);
+	GDVIRTUAL1R(int, _import_post_parse, Ref<GLTFState>);
+	GDVIRTUAL4R(int, _import_node, Ref<GLTFState>, Ref<GLTFNode>, Dictionary, Node *);
+	GDVIRTUAL2R(int, _import_post, Ref<GLTFState>, Node *);
+	GDVIRTUAL1R(int, _export_preflight, Node *);
+	GDVIRTUAL4R(int, _export_node, Ref<GLTFState>, Ref<GLTFNode>, Dictionary, Node *);
+	GDVIRTUAL1R(int, _export_post, Ref<GLTFState>);
+};
+
+#endif // GLTF_DOCUMENT_EXTENSION_H
