+++ modules/gltf/structures/gltf_buffer_view.cpp
@@ -30,8 +30,6 @@
 
 #include "gltf_buffer_view.h"
 
-#include "../gltf_document_extension.h"
-
 void GLTFBufferView::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("get_buffer"), &GLTFBufferView::get_buffer);
 	ClassDB::bind_method(D_METHOD("set_buffer", "buffer"), &GLTFBufferView::set_buffer);
